move app to react
