+++ app/assets/stylesheets/miq_v2v_ui/base.scss
@@ -1,1 +1,1 @@
-@import '../../../javascript/app/App.scss';
+@import '../../../javascript/react/App.scss';
