+++ app/javascript/react/config/Toolbar.js
