+++ app/javascript/react/index.js
