+++ app/javascript/react/screens/App/App.scss
