+++ app/javascript/react/screens/App/Dashboard/__test__/Dashboard.test.js
