+++ app/javascript/react/screens/App/Overview/components/ModalWizard/ModalWizardBody.js
