+++ app/javascript/react/screens/App/Overview/screens/MappingWizard/MappingWizardActions.js
