+++ app/javascript/react/screens/App/Overview/screens/MappingWizard/MappingWizardBody.js
