+++ app/javascript/react/screens/App/Overview/screens/MappingWizard/MappingWizardSelectors.js
