+++ app/javascript/react/screens/App/Overview/screens/MappingWizard/__tests__/MappingWizard.test.js
