+++ app/javascript/react/screens/App/Overview/screens/MappingWizard/__tests__/index.test.js
