+++ app/javascript/react/screens/App/Overview/screens/MappingWizard/components/MappingWizardClustersStep/mappingWizardClustersStep.fixtures.js
