+++ app/javascript/redux/reducers/index.js
@@ -1,7 +1,7 @@
 import { combineReducers } from 'redux';
-import { reducers as overviewReducers } from '../../app/screens/App/Overview';
-import { reducers as mappingWizardReducers } from '../../app/screens/App/Overview/screens/MappingWizard';
-import { reducers as mappingWizardClustersStepReducers } from '../../app/screens/App/Overview/screens/MappingWizard/components/MappingWizardClustersStep';
+import { reducers as overviewReducers } from '../../react/screens/App/Overview';
+import { reducers as mappingWizardReducers } from '../../react/screens/App/Overview/screens/MappingWizard';
+import { reducers as mappingWizardClustersStepReducers } from '../../react/screens/App/Overview/screens/MappingWizard/components/MappingWizardClustersStep';
 
 export function combineReducersAsync(asyncReducers) {
   return combineReducers({
